complete multichoice
